update quyen user trong all_controller

+++ app/Http/Controllers/Paper_DutuController.php
@@ -95,6 +95,11 @@ class Paper_DutuController extends Controller
     public function destroy($id)
     {
         //
+        if(Auth::user()->roleid!=1)
+        {
+            return Redirect::back()->with('message','Bạn không có quyền thực hiện hành động này!!!');
+        }
 		PaperDutu::where('id',$id)->delete();
+        return Redirect::back();
     }
 }
