+++ app/Http/Controllers/PostController.php
@@ -90,6 +90,11 @@ class PostController extends Controller
     public function destroy($id)
     {
         //
+        if(Auth::user()->roleid!=1)
+        {
+            return Redirect::back()->with('message','Bạn không có quyền thực hiện hành động này!!!');
+        }
 		Post::where('id',$id)->delete();
+        return Redirect::back();
     }
 }
