+++ app/Http/Controllers/RoleController.php
@@ -91,6 +91,11 @@ class RoleController extends Controller
     public function destroy($id)
     {
         //
+        if(Auth::user()->roleid!=1)
+        {
+            return Redirect::back()->with('message','Bạn không có quyền thực hiện hành động này!!!');
+        }
 		Role::where('id',$id)->delete();
+        return Redirect::back();
     }
 }
